pyproject.toml
[project]
name = "resume-extraction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "google-genai>=1.59.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
]
